backend/requirements.txt
fastapi
uvicorn
osmnx
networkx
rasterio
numpy
scipy
shapely
geopy
pydantic
pillow